Fix email deliveries foreign key type

+++ database/migrations/2026_03_01_000000_cleanup_and_create_email_deliveries.php
@@ -24,7 +24,7 @@ return new class extends Migration
                 $table->string('template_key')->nullable();
                 $table->string('recipient');
                 $table->string('recipient_email')->nullable();
-                $table->unsignedBigInteger('user_id')->nullable();
+                $table->unsignedInteger('user_id')->nullable();
                 $table->string('subject');
                 $table->string('status')->default('queued'); // queued, sent, failed, deferred, blocked, skipped (validated in application layer, see EmailDelivery status constants)
                 $table->string('provider')->nullable()->default('resend');
@@ -72,7 +72,7 @@ return new class extends Migration
                 $table->string('recipient_email')->nullable()->after('recipient');
             }
             if (!Schema::hasColumn('email_deliveries', 'user_id')) {
-                $table->unsignedBigInteger('user_id')->nullable()->after('recipient_email');
+                $table->unsignedInteger('user_id')->nullable()->after('recipient_email');
             }
             if (!Schema::hasColumn('email_deliveries', 'subject')) {
                 $table->string('subject')->after('user_id');
